+++ demo/atlas.test.ts
@@ -1,10 +1,10 @@
-import { AtlasMeta } from '@/atlas-pack'
+import { Atlas } from '@/atlas-pack'
 import { assertEquals } from 'std/testing/asserts.ts'
 import atlasJSON from './atlas.json' assert { type: 'json' }
 
-Deno.test('AtlasMeta.toJSON()', () => {
+Deno.test('Atlas.toJSON()', () => {
   assertEquals(
     atlasJSON,
-    JSON.parse(JSON.stringify(AtlasMeta.fromJSON(atlasJSON))),
+    JSON.parse(JSON.stringify(Atlas.fromJSON(atlasJSON))),
   )
 })