+++ mod.ts
@@ -1,5 +1,5 @@
 export * from './src/animator.ts'
 export * from './src/aseprite.ts'
-export * from './src/atlas-meta-parser.ts'
-export * from './src/atlas-meta.ts'
+export * from './src/atlas-parser.ts'
+export * from './src/atlas.ts'
 export * from './src/film.ts'