+++ docs/changelog.md
@@ -5,6 +5,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 ## v8.0.0
 
 - \[API\] Drop namespaces and remove int typing.
+- \[API\] Rename AtlasMeta to Atlas.
 
 ## v7.1.0